Improved approach
Improved the approach, instead of proving the mean-field equations from the start we consider the cumulant approximation (for one operator it is the same as the exact case) and we obtain the mean-field equation using <a_ib_j>=<a_i><b_j> with i!=j. Also, I introduced a more general interaction Hamiltonian which takes 4 indexes, two indexes related to the site and two indexes related to the noise operator. Double-check the algebra until equation 65.